+++ include/pose.h
@@ -9,12 +9,12 @@ class Pose
 {
 public:
   Pose();
-  Pose(long aX, long aY, int orientation, int range, double FOV);
+  Pose(float aX, float aY, float orientation, int range, double FOV);
   virtual ~Pose();
   double getDistance( Pose &pose);   
-  long getX();
-  long getY();
-  int getOrientation();
+  float getX();
+  float getY();
+  float getOrientation();
   int getRange();
   double getFOV();
   bool isEqual(Pose &p);
@@ -26,8 +26,8 @@ public:
   
   
 protected:
-  long aX, aY;		// x and y coordinates of the cell
-  int orientation;	// orientation theta of th robot
+  float aX, aY;		// x and y coordinates of the cell
+  float orientation;	// orientation theta of th robot
   int range;		// radius of the sensing operation 
   double FOV;		// central angle of the sensing operation
   int informationGain;
